LayoutTests/css3/scroll-snap/scroll-snap-position-values-expected.txt

Tests 'initial' on scroll snap properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS noInitial.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS noInitial.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS noInitial.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS noInitial.style['-webkit-scroll-snap-destination'] is "100% 100%"
PASS window.getComputedStyle(noInitial).getPropertyValue('-webkit-scroll-snap-destination') is "100% 100%"
PASS noInitial.style['-webkit-scroll-snap-coordinate'] is "0% 0%"
PASS window.getComputedStyle(noInitial).getPropertyValue('-webkit-scroll-snap-coordinate') is "0% 0%"
PASS initialType.style['-webkit-scroll-snap-type'] is "initial"
PASS window.getComputedStyle(initialType).getPropertyValue('-webkit-scroll-snap-type') is "none"
PASS initialType.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(initialType).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS initialType.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(initialType).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS initialType.style['-webkit-scroll-snap-destination'] is "0% 0%"
PASS window.getComputedStyle(initialType).getPropertyValue('-webkit-scroll-snap-destination') is "0% 0%"
PASS initialType.style['-webkit-scroll-snap-coordinate'] is "100% 100%"
PASS window.getComputedStyle(initialType).getPropertyValue('-webkit-scroll-snap-coordinate') is "100% 100%"
PASS initialXPoints.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(initialXPoints).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS initialXPoints.style['-webkit-scroll-snap-points-x'] is "initial"
PASS window.getComputedStyle(initialXPoints).getPropertyValue('-webkit-scroll-snap-points-x') is "none"
PASS initialXPoints.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(initialXPoints).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS initialXPoints.style['-webkit-scroll-snap-destination'] is "50% 50%"
PASS window.getComputedStyle(initialXPoints).getPropertyValue('-webkit-scroll-snap-destination') is "50% 50%"
PASS initialXPoints.style['-webkit-scroll-snap-coordinate'] is "50% 50%"
PASS window.getComputedStyle(initialXPoints).getPropertyValue('-webkit-scroll-snap-coordinate') is "50% 50%"
PASS initialYPoints.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(initialYPoints).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS initialYPoints.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(initialYPoints).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS initialYPoints.style['-webkit-scroll-snap-points-y'] is "initial"
PASS window.getComputedStyle(initialYPoints).getPropertyValue('-webkit-scroll-snap-points-y') is "none"
PASS initialYPoints.style['-webkit-scroll-snap-destination'] is "50% 20px"
PASS window.getComputedStyle(initialYPoints).getPropertyValue('-webkit-scroll-snap-destination') is "50% 20px"
PASS initialYPoints.style['-webkit-scroll-snap-coordinate'] is "50% 50%"
PASS window.getComputedStyle(initialYPoints).getPropertyValue('-webkit-scroll-snap-coordinate') is "50% 50%"
PASS initialDestination.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(initialDestination).getPropertyValue('-webkit-scroll-snap-type') is "proximity"
PASS initialDestination.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(initialDestination).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS initialDestination.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(initialDestination).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS initialDestination.style['-webkit-scroll-snap-destination'] is "initial"
PASS window.getComputedStyle(initialDestination).getPropertyValue('-webkit-scroll-snap-destination') is "0px 0px"
PASS initialDestination.style['-webkit-scroll-snap-coordinate'] is "none"
PASS window.getComputedStyle(initialDestination).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS initialSnapCoordinate.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(initialSnapCoordinate).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS initialSnapCoordinate.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(initialSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS initialSnapCoordinate.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(initialSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS initialSnapCoordinate.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(initialSnapCoordinate).getPropertyValue('-webkit-scroll-snap-destination') is "0px 0px"
PASS initialSnapCoordinate.style['-webkit-scroll-snap-coordinate'] is "initial"
PASS window.getComputedStyle(initialSnapCoordinate).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS allInitial.style['-webkit-scroll-snap-type'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('-webkit-scroll-snap-type') is "none"
PASS allInitial.style['-webkit-scroll-snap-points-x'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('-webkit-scroll-snap-points-x') is "none"
PASS allInitial.style['-webkit-scroll-snap-points-y'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('-webkit-scroll-snap-points-y') is "none"
PASS allInitial.style['-webkit-scroll-snap-destination'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('-webkit-scroll-snap-destination') is "0px 0px"
PASS allInitial.style['-webkit-scroll-snap-coordinate'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS noneSnapCoordinateAndDestination.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(noneSnapCoordinateAndDestination).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS noneSnapCoordinateAndDestination.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(noneSnapCoordinateAndDestination).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS noneSnapCoordinateAndDestination.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(noneSnapCoordinateAndDestination).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS noneSnapCoordinateAndDestination.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(noneSnapCoordinateAndDestination).getPropertyValue('-webkit-scroll-snap-destination') is "0px 0px"
PASS noneSnapCoordinateAndDestination.style['-webkit-scroll-snap-coordinate'] is "none"
PASS window.getComputedStyle(noneSnapCoordinateAndDestination).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS calculatedCoordinateAndDestination.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(calculatedCoordinateAndDestination).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS calculatedCoordinateAndDestination.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(calculatedCoordinateAndDestination).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(96%)"
PASS calculatedCoordinateAndDestination.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(calculatedCoordinateAndDestination).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(52%)"
PASS calculatedCoordinateAndDestination.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(calculatedCoordinateAndDestination).getPropertyValue('-webkit-scroll-snap-destination') is "106px 25px"
PASS calculatedCoordinateAndDestination.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(calculatedCoordinateAndDestination).getPropertyValue('-webkit-scroll-snap-coordinate') is "16px 384px"
PASS successfullyParsed is true

TEST COMPLETE
